add management_points() entry point
All my scripts seemed to be doing the same thing - fetch
the array of gateway devices then build a dictionary. So
may as well make that a shared function.

Also, in daikin-zappi.py add the target room temperature
in addition to the current value - one goal of this is
to help tune weather curve and modulation, and for the
latter, need to know the target room temperature.